Extract preview workflow CLI command
